+++ app/src/main/java/com/lagradost/cloudstream3/ui/account/AccountSelectActivity.kt
@@ -4,15 +4,14 @@ import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.GridLayoutManager
-import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.lagradost.cloudstream3.CommonActivity
 import com.lagradost.cloudstream3.CommonActivity.loadThemes
 import com.lagradost.cloudstream3.MainActivity
 import com.lagradost.cloudstream3.R
 import com.lagradost.cloudstream3.databinding.ActivityAccountSelectBinding
-import com.lagradost.cloudstream3.databinding.ActivityAccountSelectTvBinding
-import com.lagradost.cloudstream3.ui.account.AccountDialog.showPinInputDialog
+import com.lagradost.cloudstream3.ui.account.AccountAdapter.Companion.VIEW_TYPE_EDIT_ACCOUNT
+import com.lagradost.cloudstream3.ui.account.AccountDialogs.showPinInputDialog
 import com.lagradost.cloudstream3.ui.settings.SettingsFragment.Companion.isTvSettings
 import com.lagradost.cloudstream3.utils.DataStoreHelper
 import com.lagradost.cloudstream3.utils.DataStoreHelper.getAccounts
@@ -36,24 +35,34 @@ class AccountSelectActivity : AppCompatActivity() {
 
         window.navigationBarColor = colorFromAttribute(R.attr.primaryBlackBackground)
 
-        val binding = if (isTvSettings()) {
-            ActivityAccountSelectTvBinding.inflate(layoutInflater)
-        } else ActivityAccountSelectBinding.inflate(layoutInflater)
-
+        val binding = ActivityAccountSelectBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
-        val recyclerView: RecyclerView = binding.root.findViewById(R.id.account_recycler_view)
+        val recyclerView: RecyclerView = binding.accountRecyclerView
 
-
-        val adapter = AccountAdapter(accounts) { selectedAccount ->
+        val adapter = AccountAdapter(
+            accounts,
             // Handle the selected account
-            onAccountSelected(selectedAccount)
-        }
+            accountSelectCallback = { onAccountSelected(it) },
+            // Handle the selected account
+            accountCreateCallback = { onAccountUpdated(it) },
+            accountEditCallback = { onAccountUpdated(it) }
+        )
+
         recyclerView.adapter = adapter
 
-        recyclerView.layoutManager = if (isTvSettings()) {
-            LinearLayoutManager(this)
-        } else GridLayoutManager(this, 2)
+        binding.editAccountButton.setOnClickListener {
+            adapter.viewType = VIEW_TYPE_EDIT_ACCOUNT
+            adapter.notifyDataSetChanged()
+        }
+
+        if (isTvSettings()) {
+            val spanSize = if (accounts.count() <= 6) {
+                accounts.count()
+            } else 6
+
+            recyclerView.layoutManager = GridLayoutManager(this, spanSize)
+        }
     }
 
     private fun onAccountSelected(selectedAccount: DataStoreHelper.Account) {
@@ -70,6 +79,23 @@ class AccountSelectActivity : AppCompatActivity() {
             setAccount(selectedAccount)
             navigateToMainActivity()
         }
+    }
+
+    private fun onAccountUpdated(account: DataStoreHelper.Account) {
+        val currentAccounts = DataStoreHelper.accounts.toMutableList()
+
+        val overrideIndex = currentAccounts.indexOfFirst { it.keyIndex == account.keyIndex }
+        if (overrideIndex != -1) {
+            currentAccounts[overrideIndex] = account
+        } else currentAccounts.add(account)
+
+        val currentHomePage = DataStoreHelper.currentHomePage
+        setAccount(account)
+
+        DataStoreHelper.accounts = currentAccounts.toTypedArray()
+        DataStoreHelper.currentHomePage = currentHomePage
+
+        this.recreate()
     }
 
     private fun setAccount(account: DataStoreHelper.Account) {
