+++ app/src/main/res/layout/account_edit_dialog.xml
@@ -1,4 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
+
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
@@ -28,15 +29,6 @@
             android:textSize="20sp"
             android:textStyle="bold" />
 
-        <!--        <com.google.android.material.button.MaterialButton-->
-        <!--            android:nextFocusDown="@id/repo_name_input"-->
-        <!--            android:id="@+id/list_repositories"-->
-        <!--            android:nextFocusLeft="@id/apply_btt"-->
-        <!--            android:nextFocusRight="@id/cancel_btt"-->
-        <!--            style="@style/WhiteButton"-->
-        <!--            android:layout_width="wrap_content"-->
-        <!--            android:layout_gravity="center_vertical"-->
-        <!--            android:text="@string/view_public_repositories_button_short" />-->
     </LinearLayout>
 
     <TextView
@@ -140,4 +132,5 @@
             android:nextFocusLeft="@id/apply_btt"
             android:text="@string/sort_cancel" />
     </RelativeLayout>
+
 </LinearLayout>
