+++ app/src/main/res/layout/account_list_item_add.xml
@@ -2,15 +2,14 @@
 
 <androidx.cardview.widget.CardView xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/card_view"
-    android:layout_width="100dp"
-    android:layout_height="100dp"
+    android:layout_width="110dp"
+    android:layout_height="110dp"
     android:animateLayoutChanges="true"
     android:backgroundTint="?attr/primaryGrayBackground"
-    android:foreground="?attr/selectableItemBackgroundBorderless"
+    android:foreground="?attr/selectableItemBackground"
     app:cardCornerRadius="@dimen/rounded_image_radius"
-    android:layout_margin="5dp"
+    android:layout_margin="10dp"
     android:focusable="true"
     app:layout_constraintBottom_toBottomOf="parent"
     app:layout_constraintDimensionRatio="1"
@@ -28,3 +27,4 @@
         android:contentDescription="@string/add_account" />
 
 </androidx.cardview.widget.CardView>
+
