+++ app/src/main/res/layout/account_list_item_editing.xml
@@ -4,13 +4,13 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/card_view"
-    android:layout_width="100dp"
-    android:layout_height="100dp"
+    android:layout_width="110dp"
+    android:layout_height="110dp"
     android:animateLayoutChanges="true"
     android:backgroundTint="?attr/primaryGrayBackground"
-    android:foreground="?attr/selectableItemBackgroundBorderless"
+    android:foreground="?attr/selectableItemBackground"
     app:cardCornerRadius="@dimen/rounded_image_radius"
-    android:layout_margin="5dp"
+    android:layout_margin="10dp"
     android:focusable="true"
     app:layout_constraintBottom_toBottomOf="parent"
     app:layout_constraintDimensionRatio="1"
@@ -20,10 +20,10 @@
     app:layout_constraintTop_toTopOf="parent">
 
     <ImageView
-        android:id="@+id/profile_image_background"
+        android:id="@+id/account_image"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:alpha="0.4"
+        android:alpha="0.2"
         android:contentDescription="@string/profile_background_des"
         android:scaleType="centerCrop" />
 
@@ -44,9 +44,16 @@
         android:src="@drawable/video_locked"
         android:visibility="gone" />
 
+    <ImageView
+        android:id="@+id/pencil_icon"
+        android:layout_width="24dp"
+        android:layout_height="24dp"
+        android:layout_gravity="center"
+        android:src="@drawable/ic_outline_account_circle_24"
+        android:visibility="visible" />
+
     <TextView
-        android:id="@+id/profile_text"
-        tools:text="@string/mobile_data"
+        android:id="@+id/account_name"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_gravity="center"
