+++ app/src/main/res/layout/activity_account_select.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     android:background="?attr/primaryBlackBackground"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
@@ -17,7 +18,7 @@
         android:layout_marginTop="16dp"
         android:layout_marginBottom="16dp" />
 
-    <androidx.recyclerview.widget.RecyclerView
+    <com.lagradost.cloudstream3.ui.AutofitRecyclerView
         android:id="@+id/account_recycler_view"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
@@ -25,4 +26,15 @@
         android:paddingLeft="16dp"
         android:paddingRight="16dp" />
 
+    <com.google.android.material.button.MaterialButton
+        android:id="@+id/edit_account_button"
+        style="@style/RegularButtonTV"
+
+        android:layout_width="wrap_content"
+        android:minWidth="150dp"
+
+        android:text="@string/edit"
+        app:icon="@drawable/ic_baseline_add_24" />
+
 </LinearLayout>
+
